MAINT: rename and remove files
